refactor: 🎨 extract project file list function
